+++ chats/apps/api/v1/rooms/viewsets.py
@@ -393,9 +393,7 @@ class RoomViewset(
             permission = instance.project.get_permission(request.user)
             if not permission or not permission.is_admin:
                 raise PermissionDenied(
-                    detail=_(
-                        "Agents cannot close queued rooms in this sector."
-                    ),
+                    detail=_("Agents cannot close queued rooms in this sector."),
                     code="queued_room_close_disabled",
                 )
 
@@ -433,6 +431,7 @@ class RoomViewset(
         # TODO Separate this into smaller methods
         old_instance = serializer.instance
         old_user = old_instance.user
+        old_user = old_instance.user
 
         user = self.request.data.get("user_email")
         queue = self.request.data.get("queue_uuid")
@@ -448,12 +447,11 @@ class RoomViewset(
 
         # Create transfer object based on whether it's a user or a queue transfer and add it to the history
         if user:
-            if old_instance.user is None:
-                time = timezone.now() - old_instance.modified_on
+            if old_user is None:
                 room_metric = RoomMetrics.objects.select_related("room").get_or_create(
                     room=instance
                 )[0]
-                room_metric.waiting_time += time.total_seconds()
+                room_metric.waiting_time += calculate_last_queue_waiting_time(instance)
                 room_metric.queued_count += 1
                 room_metric.save()
             else:
@@ -497,7 +495,10 @@ class RoomViewset(
         # Create a message with the transfer data and Send to the room group
         # TODO separate create message in a function
         create_room_feedback_message(
-            instance, feedback, method=RoomFeedbackMethods.ROOM_TRANSFER, requested_by=self.request.user
+            instance,
+            feedback,
+            method=RoomFeedbackMethods.ROOM_TRANSFER,
+            requested_by=self.request.user,
         )
 
         if old_user is None and user:  # queued > agent
@@ -600,7 +601,10 @@ class RoomViewset(
             "new": new_custom_field_value,
         }
         create_room_feedback_message(
-            room, feedback, method=RoomFeedbackMethods.EDIT_CUSTOM_FIELDS, requested_by=request.user
+            room,
+            feedback,
+            method=RoomFeedbackMethods.EDIT_CUSTOM_FIELDS,
+            requested_by=request.user,
         )
 
         return Response(
@@ -647,7 +651,10 @@ class RoomViewset(
             room.add_transfer_to_history(feedback)
 
             create_room_feedback_message(
-                room, feedback, method=RoomFeedbackMethods.ROOM_TRANSFER, requested_by=request.user
+                room,
+                feedback,
+                method=RoomFeedbackMethods.ROOM_TRANSFER,
+                requested_by=request.user,
             )
             room.notify_queue("update")