+++ chats/apps/csat/tasks.py
@@ -1,3 +1,10 @@
+from chats.apps.csat.flows.definitions.flow import (
+    CSAT_FLOW_DEFINITION_DATA,
+    CSAT_FLOW_VERSION,
+)
+from django.conf import settings
+from datetime import datetime, timedelta
+from chats.apps.csat.models import CSATFlowProjectConfig
 from chats.apps.projects.models.models import Project
 from chats.celery import app
 from chats.apps.rooms.models import Room
@@ -27,3 +34,34 @@ def create_csat_flow(project_uuid: str):
         cache_client=CacheClient(),
         token_generator=JWTTokenGenerator(),
     ).create_csat_flow(project)
+
+
+@app.task
+def update_all_projects_csat_flow_definition():
+    configs = CSATFlowProjectConfig.objects.filter(version__lt=CSAT_FLOW_VERSION)
+
+    expiration_time = datetime.now() + timedelta(
+        minutes=settings.CSAT_FLOW_UPDATE_EXPIRATION_TIME
+    )
+
+    for config in configs:
+        update_project_csat_flow_definition.apply_async(
+            args=[config.project.uuid, CSAT_FLOW_DEFINITION_DATA, CSAT_FLOW_VERSION],
+            expires=expiration_time,
+        )
+
+
+@app.task
+def update_project_csat_flow_definition(
+    project_uuid: str, definition: dict, version: int
+):
+    if version != CSAT_FLOW_VERSION:
+        raise ValueError("Version is not the current CSAT flow version")
+
+    project = Project.objects.get(uuid=project_uuid)
+
+    CSATFlowService(
+        flows_client=FlowRESTClient(),
+        cache_client=CacheClient(),
+        token_generator=JWTTokenGenerator(),
+    ).update_csat_flow_definition(project, definition, version)